Main

+++ SegundoModulo/Main.java
@@ -3,7 +3,7 @@ public static void main(String[] args) {
     
     String nomeDoAluno = "Sidney Gomes"; 
     
-    double media = (6.8 + 7.9 + 8.2) / 3;
+    double media = (int) (6.8 + 7.9 + 8.2) / 3;
     System.out.println("ola " + nomeDoAluno + " a sua media e " + media);
 }
 }
